pyproject.toml
[project]
name = "randy"
version = "0.0.1"
description = "Personal multi-LLM advisory committee delivered via Telegram"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=0.3",
    "httpx>=0.27",
    "python-telegram-bot>=21.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "pyright>=1.1.380",
]

[project.scripts]
randy = "randy.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/randy"]

[tool.pyright]
include = ["src", "tests", "scripts"]
extraPaths = ["src"]
pythonVersion = "3.11"
reportMissingImports = "warning"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]
